Authentication Guard + routes descriptor
