+++ src/lib/components/Header.svelte
@@ -1,19 +1,25 @@
 <script lang="ts">
     import { page } from "$app/stores";
     import { user, logout } from "$lib/stores/user";
+    import routes from "$lib/routes";
 
     $: userData = $user;
     $: items = [
-        { label: "Home", to: "/" },
-        { label: "Signin", to: "/signin", authRequired: false },
+        ...routes,
         { label: "Signout", action: logout, authRequired: true },
     ].filter((it) => {
+        if (it.forceNavigation) {
+            return true;
+        }
+
         if (!userData && it.authRequired === true) {
             return false;
         }
+
         if (userData && it.authRequired === false) {
             return false;
         }
+
         return true;
     });
 </script>
@@ -23,16 +29,14 @@
         <ul class="flex">
             {#each items as it}
                 <li class:active={$page.url.pathname === it.to}>
-                    <a
-                        on:click={it.action}
-                        class="p-2"
-                        sveltekit:prefetch
-                        href={it.to}
-                    >
+                    <a on:click={it.action} class="p-2" href={it.to}>
                         {it.label}
                     </a>
                 </li>
             {/each}
         </ul>
     </nav>
+    {#if userData}
+        <div>{userData?.email}</div>
+    {/if}
 </header>
